db_create.py
import os.path
from migrate.versioning import api
from convi.database import init_db, SQLALCHEMY_DATABASE_URI, SQLALCHEMY_MIGRATE_REPO
init_db()
if not os.path.exists(SQLALCHEMY_MIGRATE_REPO):
    api.create(SQLALCHEMY_MIGRATE_REPO, 'database repository')
    api.version_control(SQLALCHEMY_DATABASE_URI, SQLALCHEMY_MIGRATE_REPO)
else:
    api.version_control(SQLALCHEMY_DATABASE_URI, SQLALCHEMY_MIGRATE_REPO, api.version(SQLALCHEMY_MIGRATE_REPO))

    __author__ = '86286K'