requirements.txt
Flask==0.10.1
Flask-RESTful==0.3.4
Flask-SQLAlchemy==2.0
Jinja2==2.8
MarkupSafe==0.23
SQLAlchemy==1.0.8
Tempita==0.5.2
Werkzeug==0.10.4
aniso8601==1.0.0
decorator==4.0.2
itsdangerous==0.24
pbr==1.8.0
pytz==2015.4
requests==2.7.0
six==1.9.0
sqlalchemy-migrate==0.10.0
sqlparse==0.1.16
wsgiref==0.1.2
